fix(anchor): widen DataPool.buyer_count u8 → u32 (Punch #2)
settle_receipt.rs:214 used `saturating_add(1)` on a u8, capping the
on-chain buyer count at 255. The 256th settle_receipt still succeeded
(USDC pulled, BuyerSlot leaf written) but the on-chain pool's
buyer_count silently stopped advancing — divergence from reality.

Widening to u32 lifts the cap to ~4.29B receipts. PDA INIT_SPACE
grows by 3 bytes; breaking on-chain layout change so a fresh deploy
is needed (devnet redeploy already planned for the lambda field
rename earlier in this branch).

Side effect: settle_receipt's sponsor check needs an explicit cast
(`pool.buyer_count < pool.min_buyers as u32`). Tests helper signature
updated. Anchor 42/42 unchanged.

+++ anchor/programs/datapool/src/instructions/settle_receipt.rs
@@ -141,7 +141,8 @@ pub fn handle_settle_receipt<'info>(
     // timing. Decouples sponsor classification from the off-chain ordering
     // of settle_receipt vs trigger_fetch — server can call trigger_fetch
     // first and the early receipts still get is_sponsor=true.
-    let is_sponsor = ctx.accounts.pool.buyer_count < ctx.accounts.pool.min_buyers;
+    let is_sponsor =
+        ctx.accounts.pool.buyer_count < (ctx.accounts.pool.min_buyers as u32);
 
     let price = ctx.accounts.pool.current_price(now);
     require!(
